+++ src/main/java/com/yozhikovd/jm_preproject_spring_boot_rest_js_3_1_3/services/UserServiceImpl.java
@@ -27,7 +27,7 @@ public class UserServiceImpl implements UserService {
 
     @Override
     @Transactional
-    public User getUserById(int id) {
+    public User getUserById(Long id) {
         return userDao.getUserById(id);
     }
 
@@ -46,7 +46,7 @@ public class UserServiceImpl implements UserService {
 
     @Override
     @Transactional
-    public void deleteUser(int id) {
+    public void deleteUser(Long id) {
         userDao.deleteUser(id);
     }
 
@@ -76,7 +76,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-      public UserDto getUserFromID(int id) {
+      public UserDto getUserFromID(Long id) {
 
       return convertToUserDTO(getUserById((id)));
